Split Position internals into their own file
